segmeting classes
