+++ entities/bullet_entity.py
@@ -1,0 +1,15 @@
+class Bullet(sprite.Sprite):
+    def __init__(self, xpos, ypos, direction, speed, filename, side):
+        sprite.Sprite.__init__(self)
+        self.image = IMAGES[filename]
+        self.rect = self.image.get_rect(topleft=(xpos, ypos))
+        self.speed = speed
+        self.direction = direction
+        self.side = side
+        self.filename = filename
+
+    def update(self, keys, *args):
+        game.screen.blit(self.image, self.rect)
+        self.rect.y += self.speed * self.direction
+        if self.rect.y < 15 or self.rect.y > 600:
+            self.kill()
