+++ paths.py
@@ -1,4 +1,3 @@
-
 from os.path import abspath, dirname
