+++ util/images_render.py
@@ -1,9 +1,8 @@
-from paths import *
 from pygame import *
 
 IMG_NAMES = ['player',
              'laser', ]
-
+IMAGE_PATH = "C:\Users\esdra\Documents\ip_crew\crew\images/"
 
 IMAGES = {name: image.load(IMAGE_PATH + '{}.png'.format(name)).convert_alpha()
           for name in IMG_NAMES}
